+++ traitlets/tests/utils.py
@@ -1,8 +1,10 @@
+from __future__ import annotations
+
 import sys
 from subprocess import PIPE, Popen
 
 
-def get_output_error_code(cmd):
+def get_output_error_code(cmd: str) -> tuple[str, str, int]:
     """Get stdout, stderr, and exit code from running a command"""
     p = Popen(cmd, stdout=PIPE, stderr=PIPE)  # noqa
     out, err = p.communicate()
@@ -11,7 +13,7 @@ def get_output_error_code(cmd):
     return out, err, p.returncode
 
 
-def check_help_output(pkg, subcommand=None):
+def check_help_output(pkg: str, subcommand: str | None = None) -> tuple[str, str]:
     """test that `python -m PKG [subcommand] -h` works"""
     cmd = [sys.executable, "-m", pkg]
     if subcommand:
@@ -25,7 +27,7 @@ def check_help_output(pkg, subcommand=None):
     return out, err
 
 
-def check_help_all_output(pkg, subcommand=None):
+def check_help_all_output(pkg: str, subcommand: str | None = None) -> tuple[str, str]:
     """test that `python -m PKG --help-all` works"""
     cmd = [sys.executable, "-m", pkg]
     if subcommand:
